+++ nameko/click_cli/click_options.py
@@ -1,6 +1,6 @@
-"""Custom click option and argument decorators
+"""Custom click option decorators
 
-Allow reuse of options and arguments across subcommands.
+Allow reuse of options across subcommands.
 
 Reuse in other click based commands is also possible.
 
@@ -10,7 +10,7 @@ from functools import partial
 
 import click
 
-from .paramtypes import HOST_PORT, KEY_VAL, NamekoModuleServicesParamType
+from .click_paramtypes import HOST_PORT, KEY_VAL
 
 
 def flatten_list_of_lists(ctx, param, list_of_lists):
@@ -55,14 +55,4 @@ option_backdoor_port = partial(
 )
 option_interface = partial(
     click.option, "--interface", type=click.Choice(["bpython", "ipython", "plain"])
-)
-argument_services = partial(
-    click.argument,
-    "services",
-    nargs=-1,
-    required=True,
-    metavar="module[:service class]",
-    # allow import of modules from current dir (manipulate sys.path)
-    type=NamekoModuleServicesParamType(extra_sys_paths=["."]),
-    callback=flatten_list_of_lists,
 )
