+++ nameko/click_cli/do_shell.py
