--- nameko/click_cli/show_config.py
@@ -1,9 +1,0 @@
-from __future__ import print_function
-
-import yaml
-
-from nameko import config
-
-
-def main(args):
-    print(yaml.dump(config.data))
