+++ test/click_cli/test_cmain.py
@@ -42,7 +42,7 @@ def test_run():
     config_file = None
     broker = None
     backdoor_port = None
-    with patch("nameko.click_cli.run.main") as main_run:
+    with patch("nameko.click_cli.do_run.main") as main_run:
         with patch("nameko.click_cli.setup_config") as setup_config:
             cli(standalone_mode=False)
             assert setup_config.call_args_list == [call(config_file, define, broker)]
@@ -51,7 +51,7 @@ def test_run():
 
 @pytest.mark.parametrize("exception", (CommandError, ConfigurationError))
 def test_error(exception, capsys):
-    with patch("nameko.click_cli.run.main") as main_run:
+    with patch("nameko.click_cli.do_run.main") as main_run:
         with pytest.raises(SystemExit):
             main_run.side_effect = exception("boom")
             cli(standalone_mode=True)
@@ -82,7 +82,7 @@ def test_error(exception, capsys):
     ),
 )
 def test_parse_config_option(text, expected_key, expected_value):
-    from nameko.click_cli.myclick.paramtypes import KeyValParamType
+    from nameko.click_cli.click_paramtypes import KeyValParamType
 
     assert KeyValParamType().convert(text, None, None) == (expected_key, expected_value)
